pkg/shortener/base62/base62.go
package base62

import (
	"errors"
	"math"
	"strings"

	"github.com/VladimirStepanov/urlshortener/pkg/shortener"
)

const (
	alphabet = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789"
	length   = uint64(len(alphabet))
)

//Base62 ...
type Base62 struct{}

//New -  create new Base62 Shortener
func New() shortener.Shortener {
	return &Base62{}
}

//Encode ...
func (b *Base62) Encode(number uint64) string {
	var encodedBuilder strings.Builder
	encodedBuilder.Grow(11)

	for ; number > 0; number = number / length {
		encodedBuilder.WriteByte(alphabet[(number % length)])
	}

	return encodedBuilder.String()
}

//Decode ...
func (b *Base62) Decode(encoded string) (uint64, error) {
	var number uint64

	for i, symbol := range encoded {
		alphabeticPosition := strings.IndexRune(alphabet, symbol)

		if alphabeticPosition == -1 {
			return uint64(alphabeticPosition), errors.New("invalid character: " + string(symbol))
		}
		number += uint64(alphabeticPosition) * uint64(math.Pow(float64(length), float64(i)))
	}

	return number, nil
}
